+++ example/lib/main.dart
@@ -1,8 +1,8 @@
-import 'dart:convert';
-
 import 'package:flutter/material.dart';
 import 'package:square_web_payments/square_web_payments.dart';
 import 'package:widgetbook/widgetbook.dart';
+
+import 'payment_builder.dart';
 
 final Payments payments =
     Square.payments('sandbox-sq0idb-negdp6Z5Q7RYeFbAv38m0A', 'L6XCYJM4ZPHRP');
@@ -18,78 +18,44 @@ class WidgetbookApp extends StatelessWidget {
   Widget build(BuildContext context) => Widgetbook.material(directories: [
         WidgetbookUseCase(
             name: 'Apple Pay',
-            builder: (context) => _buildPaymentMethod(
-                payments.applePay(payments.paymentRequest(
+            builder: (context) => PaymentBuilder(
+                future: payments.applePay(payments.paymentRequest(
                     const PaymentRequestOptions(
                         countryCode: 'US',
                         currencyCode: 'USD',
                         total: LineItem(amount: '1.00', label: 'Total')))),
-                (applePay) => ApplePayView(
-                    applePay: applePay,
-                    onPressed: () => _tokenize(context, applePay)))),
+                builder: (applePay, tokenize) =>
+                    ApplePayView(applePay: applePay, onPressed: tokenize))),
         WidgetbookUseCase(
             name: 'Card',
-            builder: (context) => _buildPaymentMethod(
-                payments.card(),
-                (card) => Column(children: [
+            builder: (context) => PaymentBuilder(
+                future: payments.card(),
+                builder: (card, tokenize) => Column(children: [
                       CardView(card: card),
                       TextButton(
-                          onPressed: () => _tokenize(context, card),
-                          child: const Text('Tokenize'))
+                          onPressed: tokenize, child: const Text('Tokenize'))
                     ]))),
         WidgetbookUseCase(
             name: 'Gift Card',
-            builder: (context) => _buildPaymentMethod(
-                payments.giftCard(),
-                (card) => Column(children: [
-                      CardView(card: card),
+            builder: (context) => PaymentBuilder(
+                future: payments.giftCard(),
+                builder: (giftCard, tokenize) => Column(children: [
+                      CardView(card: giftCard),
                       TextButton(
-                          onPressed: () => _tokenize(context, card),
-                          child: const Text('Tokenize'))
+                          onPressed: tokenize, child: const Text('Tokenize'))
                     ]))),
         WidgetbookUseCase(
             name: 'Google Pay',
-            builder: (context) => _buildPaymentMethod(
-                payments.googlePay(payments.paymentRequest(
+            builder: (context) => PaymentBuilder(
+                future: payments.googlePay(payments.paymentRequest(
                     const PaymentRequestOptions(
                         countryCode: 'US',
                         currencyCode: 'USD',
                         total: LineItem(amount: '1.00', label: 'Total')))),
-                (googlePay) => GooglePayView(
+                builder: (googlePay, tokenize) => GooglePayView(
                     googlePay: googlePay,
                     googlePayButtonOptions:
                         const GooglePayButtonOptions(buttonSizeMode: 'fill'),
-                    onPressed: () => _tokenize(context, googlePay)))),
+                    onPressed: tokenize))),
       ]);
-
-  Widget _buildPaymentMethod<TPaymentMethod extends PaymentMethod>(
-          Future<TPaymentMethod> future,
-          Widget Function(TPaymentMethod paymentMethod) builder) =>
-      Container(
-          constraints: const BoxConstraints.expand(),
-          decoration: const BoxDecoration(color: Colors.white),
-          padding: const EdgeInsets.all(8),
-          child: FutureBuilder(
-              future: future,
-              builder: (context, snapshot) =>
-                  snapshot.connectionState == ConnectionState.done
-                      ? snapshot.data != null
-                          ? builder(snapshot.data!)
-                          : SelectableText(const JsonEncoder.withIndent('  ')
-                              .convert(snapshot.error))
-                      : const Center(child: CircularProgressIndicator())));
-
-  void _tokenize(BuildContext context, PaymentMethod paymentMethod) =>
-      showDialog(
-          context: context,
-          builder: (BuildContext context) => FutureBuilder(
-              future: paymentMethod.tokenize(),
-              builder: (context, snapshot) => AlertDialog(
-                  title: snapshot.connectionState == ConnectionState.done
-                      ? Text(snapshot.hasData ? 'TokenResult' : 'Error')
-                      : const Center(child: CircularProgressIndicator()),
-                  content: snapshot.connectionState == ConnectionState.done
-                      ? SelectableText(const JsonEncoder.withIndent('  ')
-                          .convert(snapshot.data ?? snapshot.error))
-                      : null)));
 }