+++ CHANGELOG.md
@@ -36,3 +36,8 @@
 
 * Adding inline documentation for widgets
 * Adding support for Google Pay
+
+## 0.0.10
+
+* Renaming Card to PaymentCard to avoid collision with material class
+* Detaching instead of destroying payment method when view is disposed